+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/state/messages/attachments/AttachmentState.kt
@@ -26,13 +26,44 @@ import io.getstream.chat.android.compose.state.mediagallerypreview.MediaGalleryP
  * @param message Data that represents the message information.
  * @param onLongItemClick Handler for a long click on the message item.
  * @param onImagePreviewResult Handler when the user selects an action to scroll to and focus an image.
- * @param onMediaPreviewResult Handler used when the user selects an action to perform from
+ * @param onMediaGalleryPreviewResult Handler used when the user selects an action to perform from
  * [io.getstream.chat.android.compose.ui.attachments.preview.MediaGalleryPreviewActivity].
  */
-public data class AttachmentState(
+public data class AttachmentState
+@Deprecated(
+    message = "Constructor containing parameter 'onImagePreviewResult' has been deprecated, " +
+        "please use the constructor that does not have said parameter. Parameter 'onMediaPreviewResult' " +
+        "is the replacement for the parameter 'onImagePreviewResult' and is functionally the same.",
+    level = DeprecationLevel.WARNING
+)
+constructor(
     val message: Message,
     val onLongItemClick: (Message) -> Unit = {},
+    @Deprecated(
+        message = "This parameter has been deprecated. Replace it with" +
+            "'AttachmentState.onMediaPreview'.",
+        level = DeprecationLevel.WARNING
+    )
     val onImagePreviewResult: (ImagePreviewResult?) -> Unit = {},
-    // TODO structure better with a deprecation process
-    val onMediaPreviewResult: (MediaGalleryPreviewResult?) -> Unit
-)
+    val onMediaGalleryPreviewResult: (MediaGalleryPreviewResult?) -> Unit,
+) {
+
+    /**
+     * Represents the state of Attachment items, used to render and add handlers required for the attachment to work.
+     *
+     * @param message Data that represents the message information.
+     * @param onLongItemClick Handler for a long click on the message item.
+     * @param onMediaGalleryPreviewResult Handler used when the user selects an action to perform from
+     * [io.getstream.chat.android.compose.ui.attachments.preview.MediaGalleryPreviewActivity].
+     */
+    public constructor(
+        message: Message,
+        onLongItemClick: (Message) -> Unit = {},
+        onMediaGalleryPreviewResult: (MediaGalleryPreviewResult?) -> Unit,
+    ) : this(
+        message = message,
+        onLongItemClick = onLongItemClick,
+        onImagePreviewResult = {},
+        onMediaGalleryPreviewResult = onMediaGalleryPreviewResult,
+    )
+}