+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/messages/GiphyMessageContent.kt
@@ -102,7 +102,7 @@ public fun GiphyMessageContent(
         MessageAttachmentsContent(
             message = message,
             onLongItemClick = {},
-            onImagePreviewResult = {},
+            onMediaGalleryPreviewResult = {},
         )
 
         Spacer(