+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/mediapreview/MediaGalleryPreviewViewModelFactory.kt
@@ -19,13 +19,21 @@ package io.getstream.chat.android.compose.viewmodel.mediapreview
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.getstream.chat.android.client.ChatClient
+import io.getstream.chat.android.client.setup.state.ClientState
 
 /**
  * Holds the dependencies required for the Media Preview Screen.
  * Currently builds the [MediaGalleryPreviewViewModel] using those dependencies.
+ *
+ * @param chatClient An instance of the low level client
+ * used for basic chat API functionality.
+ * @param clientState Holds information about the current SDK state.
+ * @param messageId The ID of the message we are fetching in order
+ * to display the attachments.
  */
 public class MediaGalleryPreviewViewModelFactory(
     private val chatClient: ChatClient,
+    private val clientState: ClientState = chatClient.clientState,
     private val messageId: String,
 ) : ViewModelProvider.Factory {
 
@@ -34,6 +42,10 @@ public class MediaGalleryPreviewViewModelFactory(
      */
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         @Suppress("UNCHECKED_CAST")
-        return MediaGalleryPreviewViewModel(chatClient, messageId) as T
+        return MediaGalleryPreviewViewModel(
+            chatClient = chatClient,
+            clientState = clientState,
+            messageId = messageId
+        ) as T
     }
 }